refactor(specs): ADR-100 Phase 10+ — PixiRenderer.ts 완전 제거
ADR-100 Phase 9에서 Builder Canvas의 PixiJS 런타임은 제거됐으나
packages/specs/src/renderers/PixiRenderer.ts가 dead code로 잔존.
renderToPixi() 호출자 0건이며, 내부 헬퍼 getVariantColors/getSizePreset만
Skia 경로(useSpecRenderer, elementHelpers)에서 재사용 중이었음.

변경:
- renderers/utils/variantColors.ts 신설 — getVariantColors + getSizePreset 이전
  (pixi 의존 없는 순수 함수, VariantSpec/SizeSpec 타입만 사용)
- renderers/PixiRenderer.ts 삭제 (-289 LOC, renderToPixi + PixiRenderContext)
- renderers/index.ts — PixiRenderer export 제거, variantColors 재수출
- src/index.ts — PixiRenderContext / renderToPixi export 제거
- packages/specs/package.json — pixi.js peer dep 제거 (+ peerDependenciesMeta)

검증:
- grep renderToPixi|PixiRenderContext → 0 hits
- pnpm type-check 3/3 PASS
- pnpm build:specs — 107 CSS 파일 생성 정상
