+++ packages/specs/package.json
@@ -63,14 +63,10 @@
     "vitest": "^4.0.18"
   },
   "peerDependencies": {
-    "pixi.js": "^8.16.0",
     "react": "^18.0.0 || ^19.0.0",
     "lucide-react": "^0.575.0"
   },
   "peerDependenciesMeta": {
-    "pixi.js": {
-      "optional": true
-    },
     "react": {
       "optional": true
     },
